rename: dto에서 validator로 이름 수정
