+++ springbasic/src/main/java/com/programmers/springbasic/domain/voucher/validator/VoucherCommandValidator.java
@@ -1,16 +1,16 @@
-package com.programmers.springbasic.domain.voucher.dto.request;
+package com.programmers.springbasic.domain.voucher.validator;
 
 import lombok.Getter;
 
 
 @Getter
-public class CommandDTO {
-    private static final String VALID_COMMAND_REGEXP = "^(exit|create|list)$";  // exit, create, list 허용
+public class VoucherCommandValidator {
+    private static final String VALID_COMMAND_REGEXP = "^(exit|create|list|read|update|delete|show)$";  // exit, create, list 허용
     private static final String INVALID_COMMAND_MESSAGE = "지원하지 않는 command입니다.";
 
     private String inputCommand;
 
-    public CommandDTO(String inputCommand) {
+    public VoucherCommandValidator(String inputCommand) {
         if (!inputCommand.matches(VALID_COMMAND_REGEXP)) {
             throw new IllegalArgumentException(INVALID_COMMAND_MESSAGE);
         }
